Change data update behaviour to use rsync
This uses rsync to update the data in a user's home directory. With this
we change the behaviour so that:
* if a user edited a data file it will not be overwritten
* if there are new data files in the "source" they will be copied over

What we do not support right now is a file in the source being modified
and then overwrite what the user has in their hme directory. For files
that already exist in the user's home directory that version always
wins. In order to get a new version a user has to explicitly delete the
file they want to update in their home directory, stop their server,
wait for it to stop, start the server again.